refactor(#78): 서비스 레이어 분리
